requirements.txt
transformers==4.55.4
datasets==4.0.0
peft==0.17.1
accelerate==1.10.0
bitsandbytes==0.47.0
pyyaml==6.0.2         
rouge-score==0.1.2    
bert-score==0.3.13     
nltk==3.9.1           

